Update udemy_enroller_vanilla.py
After the first time, I don't need it to go through all the pages of tutorialbar.com. Most of the coupons past page 5 to 10 are expired or no good. I put this in to only go through the first 6 pages and then stop. This was a quick edit, it would probably be better to put this in the settings.txt.