Fix a bug where we try to access an invalid page from Udemy which raises an exception
For some reason, the for loop is up to total_pages +2 and not total_pages +1.
Instead of going this route, I saw that the response contains a "next" key which is None when there is no next page to load, so I replaced the for loop with a while loop which fixed my issue.